[style] integration / 共通関数file,response file,getのtest case fileに分けた
