+++ test/webserv/integration/test_http_get.py
@@ -1,37 +1,11 @@
-import os
-import subprocess
-
 import pytest
-from common import *
+from common_functions import send_request_and_assert_response
+from common_response import *
 
-
-def build_client_module():
-    # client_module ディレクトリに移動
-    client_module_dir = os.path.join(os.path.dirname(__file__), "client_module")
-    # コマンドを実行
-    subprocess.run(
-        ["python3", "setup.py", "build_ext", "--inplace"],
-        cwd=client_module_dir,
-        check=True,
-    )
-
-
-try:
-    # client_module をビルド
-    build_client_module()
-    # client_module.client をインポート
-    import client_module.client as client
-except subprocess.CalledProcessError as e:
-    print(f"Build failed: {e}")
-except ImportError as e:
-    print(f"Import failed: {e}")
-
-
-def send_request_and_assert_response(request_file, expected_response):
-    client_instance = client.Client(8080)
-    request, _ = read_file_binary(request_file)
-    response = client_instance.SendRequestAndReceiveResponse(request)
-    assert_response(response, expected_response)
+REQUEST_DIR = "test/common/request/"
+REQUEST_GET_2XX_DIR = REQUEST_DIR + "get/2xx/"
+REQUEST_GET_4XX_DIR = REQUEST_DIR + "get/4xx/"
+REQUEST_GET_5XX_DIR = REQUEST_DIR + "get/5xx/"
 
 
 @pytest.mark.parametrize(
